Use iso-8601 dates at all times.
Drop usage of JS Date in order to simplify all operations.

Extract some parts of SAView into separate component files.
