+++ frontend/src/components/manage/Blocks/Replenishment/SAView/SAView.tsx
@@ -13,35 +13,30 @@ import {
 } from 'react'
 
 import Big from 'big.js'
-import { reverse } from 'lodash'
-import { useSnackbar } from 'notistack'
 
 import DownloadButtons from '@ors/app/replenishment/DownloadButtons'
-import { AddButton, SubmitButton } from '@ors/components/ui/Button/Button'
+import { DateRangeInput } from '@ors/components/manage/Blocks/Replenishment/Inputs/DateRangeInput'
+import COLUMNS from '@ors/components/manage/Blocks/Replenishment/SAView/COLUMNS'
+import { SaveManager } from '@ors/components/manage/Blocks/Replenishment/SAView/SaveManager'
+import {
+  asDecimal,
+  formatIso8601DateString,
+} from '@ors/components/manage/Blocks/Replenishment/utils'
+import { AddButton } from '@ors/components/ui/Button/Button'
 import Link from '@ors/components/ui/Link/Link'
 import ReplenishmentContext from '@ors/contexts/Replenishment/ReplenishmentContext'
 import SoAContext from '@ors/contexts/Replenishment/SoAContext'
-import { api, formatApiUrl, getFloat } from '@ors/helpers'
-import { useStore } from '@ors/store'
+import { formatApiUrl, getFloat } from '@ors/helpers'
 
-import FormEditDialog from '../FormEditDialog'
-import { DateInput, FormattedNumberInput, Input } from '../Inputs'
-import PopoverInput from '../StatusOfTheFund/editDialogs/PopoverInput'
+import { FormattedNumberInput } from '../Inputs'
 import { SortDirection } from '../Table/types'
-import { asDecimal, dateForInput, dateFromInput } from '../utils'
 import SATable from './SATable'
 import {
   SAContribution,
   SAContributionForSave,
-  SAViewWrapperProps,
-  SaveData,
-  SaveManagerProps,
-} from './types'
-import {
-  DateRangeInputProps,
-  FileForUpload,
   SATableColumn,
   SAViewProps,
+  SAViewWrapperProps,
 } from './types'
 import {
   checkQualifiesForFerm,
@@ -52,113 +47,6 @@ import {
   sortSATableData,
   sumColumns,
 } from './utils'
-
-export function encodeFileForUpload(file: File) {
-  function resolver(resolve: (value: FileForUpload) => void) {
-    const r = new FileReader()
-    r.onload = function (evt) {
-      const read = evt.target?.result
-      if (read) {
-        resolve({
-          contentType: file.type,
-          data: (read as string).split(',')[1],
-          encoding: 'base64',
-          filename: file.name,
-        })
-      }
-    }
-    r.readAsDataURL(file)
-  }
-  return new Promise<FileForUpload>(resolver)
-}
-
-const COLUMNS: SATableColumn[] = [
-  { field: 'country', label: 'Country' },
-  {
-    editable: true,
-    field: 'un_soa',
-    label: 'UN scale of assessment *',
-    parser: asDecimal,
-    subLabel: '( [UN_SCALE_PERIOD] )',
-  },
-  {
-    confirmationText:
-      'If you make this change, this value will be fixed and all other values except for the USA will be changed accordingly.',
-    editable: true,
-    field: 'adj_un_soa',
-    label: 'Adjusted UN Scale of Assessment',
-    parser: asDecimal,
-    validator: function (value) {
-      if (value > 22) {
-        return "Value can't be greater than 22."
-      }
-    },
-  },
-  {
-    field: 'annual_contributions',
-    label: 'Annual contributions',
-    subLabel: '([PERIOD] in USD)',
-  },
-  {
-    editable: true,
-    field: 'avg_ir',
-    label: 'Average inflation rate **',
-    parser: asDecimal,
-    subLabel: '( [PREV_PERIOD] )',
-  },
-  {
-    editOptions: [
-      { label: 'Yes', value: 'true' },
-      { label: 'No', value: 'false' },
-    ],
-    editParser: function (v) {
-      return v ? v.toString() : 'false'
-    },
-    // editWidget: 'select',
-    // editable: true,
-    field: 'qual_ferm',
-    label: 'Qualifying for fixed exchange rate mechanism',
-    parser: function (v) {
-      return v === 'true' || v === 't' || v === 'y' || v === '1'
-    },
-    subLabel: '(Yes / No)',
-  },
-  {
-    className: 'print:hidden',
-    editOptions: [
-      { label: 'Yes', value: 'true' },
-      { label: 'No', value: 'false' },
-    ],
-    editParser: function (v) {
-      return v.toString()
-    },
-    editWidget: 'select',
-    editable: true,
-    field: 'opted_for_ferm',
-    label: 'Opted for fixed exchange rate mechanism',
-    parser: function (v) {
-      return v === 'true' || v === 't' || v === 'y' || v === '1'
-    },
-    subLabel: '(Yes / No)',
-  },
-  {
-    editable: true,
-    field: 'ferm_rate',
-    label: 'Currency rate of exchange used for fixed exchange ***',
-    parser: asDecimal,
-    subLabel: '[DATE_RANGE]',
-  },
-  {
-    editable: true,
-    field: 'ferm_cur',
-    label: 'National currency used for fixed exchange',
-  },
-  {
-    field: 'ferm_cur_amount',
-    label: 'Contribution amount in national currencies',
-    subLabel: '(for fixed exchange mechanism)',
-  },
-]
 
 function getEditableFieldNames(cs: SATableColumn[]) {
   const r = []
@@ -171,250 +59,6 @@ function getEditableFieldNames(cs: SATableColumn[]) {
 }
 
 const EDITABLE = getEditableFieldNames(COLUMNS)
-
-function SaveManager(props: SaveManagerProps) {
-  const { comment, currencyDateRange, data, replenishment, version, versions } =
-    props
-
-  const projectSlice = useStore((state) => state.projects)
-  const meetings = projectSlice.meetings.data
-  const formattedMeetings = meetings?.map((meeting: any) => ({
-    label: meeting.number,
-    value: meeting.id,
-    year: meeting.date ? new Date(meeting.date).getFullYear() : '-',
-  }))
-  const meetingOptions = reverse(formattedMeetings)
-
-  const { refetchData: refetchReplenishment } = useContext(ReplenishmentContext)
-  const { refetchData: refetchSoA, setCurrentVersion } = useContext(SoAContext)
-  const ctx = useContext(ReplenishmentContext)
-
-  const [isFinal, setIsFinal] = useState(false)
-  const [createNewVersion, setCreateNewVersion] = useState(true)
-  const [saving, setSaving] = useState(false)
-
-  const { enqueueSnackbar } = useSnackbar()
-
-  useEffect(
-    function () {
-      setIsFinal(version?.is_final ?? false)
-    },
-    [version],
-  )
-
-  function handleChangeFinal() {
-    setIsFinal(function (prev) {
-      return !prev
-    })
-  }
-
-  function handleChangeCreateNewVersion() {
-    setCreateNewVersion(function (prev) {
-      return !prev
-    })
-  }
-
-  function handleSave() {
-    setSaving(true)
-  }
-
-  async function confirmSave(formData: FormData) {
-    const saveData: SaveData = {
-      ...Object.fromEntries(formData.entries()),
-      amount: new Big(replenishment.amount).toString(),
-      comment,
-      data,
-      replenishment_id: replenishment.id,
-    }
-
-    saveData['final'] = isFinal
-    saveData['currency_date_range_start'] =
-      currencyDateRange.start.toISOString()
-    saveData['currency_date_range_end'] = currencyDateRange.end.toISOString()
-
-    if (saveData.decision_pdf && (saveData.decision_pdf as File).size) {
-      saveData['decision_pdf'] = await encodeFileForUpload(
-        saveData.decision_pdf as File,
-      )
-    } else {
-      saveData['decision_pdf'] = null
-    }
-
-    setSaving(false)
-
-    api('api/replenishment/scales-of-assessment', {
-      data: saveData,
-      method: 'POST',
-    })
-      .then(() => {
-        refetchReplenishment()
-        refetchSoA()
-        if (createNewVersion) {
-          setCurrentVersion((prevVersion) => (prevVersion ?? 0) + 1)
-        }
-        enqueueSnackbar('Data saved successfully.', { variant: 'success' })
-      })
-      .catch((error) => {
-        error.json().then((data: Record<string, string>[]) => {
-          // Iterate over each error object and format it
-          const messages = data
-            .map((errorObj, index) => {
-              // Extract the field name and the error message
-              const fieldErrors = Object.entries(errorObj).map(
-                ([field, errors]) => {
-                  // Check if errors is an array or object
-                  const errorMessage = Array.isArray(errors)
-                    ? errors
-                        .map((error) =>
-                          typeof error === 'object'
-                            ? JSON.stringify(error)
-                            : error,
-                        )
-                        .join(' ')
-                    : typeof errors === 'object'
-                      ? JSON.stringify(errors)
-                      : errors
-
-                  return `Row ${index + 1}: field ${field} - ${errorMessage}\n`
-                },
-              )
-
-              // Join all field errors for this particular row
-              return fieldErrors.join('\n')
-            })
-            .join('\n\n') // Separate different row errors with double newlines
-
-          // Display the notification with the formatted messages
-          enqueueSnackbar(messages, {
-            style: { whiteSpace: 'pre-line' },
-            variant: 'error',
-          })
-        })
-      })
-  }
-
-  function cancelSave() {
-    setSaving(false)
-  }
-
-  const isNewestVersion = version?.version === versions[0]?.version
-
-  const showSave = !version?.is_final && isNewestVersion
-
-  return (
-    <div className="flex items-center gap-x-4 print:hidden">
-      {saving ? (
-        <FormEditDialog
-          title="Save changes?"
-          onCancel={cancelSave}
-          onSubmit={confirmSave}
-        >
-          <div className="flex justify-between gap-4">
-            <p className="w-7/12 text-lg">
-              You can specify meeting and decision numbers where this version
-              was approved.
-            </p>
-            <div className="flex w-5/12 flex-col gap-4">
-              <div className="flex gap-4">
-                <div className="flex flex-col">
-                  <label htmlFor="meeting">Meeting</label>
-                  <PopoverInput
-                    id="meeting"
-                    className="!m-0 h-12 !w-16 !py-1"
-                    clearBtnClassName="right-1"
-                    field="meeting"
-                    options={meetingOptions}
-                    placeholder="Select meeting"
-                    required={isFinal}
-                    withClear={true}
-                    withInputPlaceholder={false}
-                  />
-                </div>
-                <div className="flex flex-col">
-                  <label htmlFor="decision">Decision</label>
-                  <Input
-                    id="decision"
-                    className="!m-0 h-12 w-16 !py-1"
-                    required={isFinal}
-                    type="text"
-                  />
-                </div>
-              </div>
-              <div className="flex flex-col">
-                <label htmlFor="decision_pdf">Decision PDF</label>
-                <Input
-                  id="decision_pdf"
-                  className="!ml-0 h-10"
-                  required={false}
-                  type="file"
-                />
-              </div>
-            </div>
-          </div>
-          <div className="mt-2 flex">
-            <div className="flex items-center gap-x-2">
-              <Input
-                id="createNewVersion"
-                className="!ml-0"
-                checked={createNewVersion}
-                type="checkbox"
-                onChange={handleChangeCreateNewVersion}
-              />
-              <label htmlFor="createNewVersion">Create new version</label>
-            </div>
-          </div>
-        </FormEditDialog>
-      ) : null}
-      {showSave && ctx.isTreasurer && (
-        <>
-          <div className="flex items-center gap-x-2">
-            <Input
-              id="markAsFinal"
-              checked={isFinal}
-              type="checkbox"
-              onChange={handleChangeFinal}
-            />
-            <label htmlFor="markAsFinal">Mark as final</label>
-          </div>
-          <SubmitButton onClick={handleSave}>Save changes</SubmitButton>
-        </>
-      )}
-    </div>
-  )
-}
-
-function DateRangeInput(props: DateRangeInputProps) {
-  const { disabled, initialEnd, initialStart, onChange } = props
-
-  const [start, setStart] = useState(initialStart)
-  const [end, setEnd] = useState(initialEnd)
-
-  const handleChangeStart: ChangeEventHandler<HTMLInputElement> = (evt) => {
-    onChange(evt.target.value, end)
-    setStart(evt.target.value)
-  }
-
-  const handleChangeEnd: ChangeEventHandler<HTMLInputElement> = (evt) => {
-    onChange(start, evt.target.value)
-    setEnd(evt.target.value)
-  }
-
-  return (
-    <div className="flex">
-      <DateInput
-        disabled={disabled}
-        value={start}
-        onChange={handleChangeStart}
-      />
-      <DateInput
-        disabled={disabled}
-        min={start}
-        value={end}
-        onChange={handleChangeEnd}
-      />
-    </div>
-  )
-}
 
 function transformContributions(cs: ApiReplenishmentSoA) {
   const r: SAContribution[] = []
@@ -500,23 +144,30 @@ function getExistingCurrency(rows: SAContribution[], value: any) {
 }
 
 function getInitialCurrencyDateRange(year: number) {
-  const start = new Date(Date.UTC(year, 0, 1))
-  const end = new Date(Date.UTC(year, 6, 0))
+  const start = `${year}-01-01`
+  const end = `${year}-07-01`
   return { end, start }
 }
 
 function formatCurrencyDateRangeForHeader(dateRange: {
-  end: Date
-  start: Date
+  end: string
+  start: string
 }) {
   const { end, start } = dateRange
-  const intl = new Intl.DateTimeFormat('en-US', { month: 'short' })
-  const sameYear = end.getUTCFullYear() == start.getUTCFullYear()
-  const strStart = sameYear
-    ? `${start.getUTCDate()} ${intl.format(start)}`
-    : `${start.getUTCDate()} ${intl.format(start)} ${start.getUTCFullYear()}`
-  const strEnd = `${end.getUTCDate()} ${intl.format(end)} ${end.getUTCFullYear()}`
-  return `${strStart} - ${strEnd}`
+  let result = ''
+  if (start && end) {
+    const sameYear = start.split('-')[0] === end.split('-')[0]
+    const strStart = sameYear
+      ? formatIso8601DateString(start).split(' ').slice(0, 2).join(' ')
+      : formatIso8601DateString(start)
+    const strEnd = formatIso8601DateString(end)
+    result = `${strStart} - ${strEnd}`
+  } else if (start) {
+    result = formatIso8601DateString(start)
+  } else if (end) {
+    result = formatIso8601DateString(end)
+  }
+  return result
 }
 
 function revertAllCurrencyNames(rows: SAContribution[], value: any) {
@@ -866,8 +517,8 @@ function SAView(props: SAViewProps) {
 
   function handleChangeCurrencyDateRange(start: string, end: string) {
     setCurrencyDateRange({
-      end: dateFromInput(end),
-      start: dateFromInput(start),
+      end: end,
+      start: start,
     })
   }
 
@@ -935,8 +586,8 @@ function SAView(props: SAViewProps) {
             </label>
             <DateRangeInput
               disabled={!isEditable}
-              initialEnd={dateForInput(currencyDateRange.end)}
-              initialStart={dateForInput(currencyDateRange.start)}
+              initialEnd={currencyDateRange.end}
+              initialStart={currencyDateRange.start}
               onChange={handleChangeCurrencyDateRange}
             />
           </div>
